Define PID type as Int64
Int64 is enough to fit the actual underlying values of different platforms (Int32 on Linux, UInt32 on Windows).

This will allow to make Process cross-platform without returning a different type to users on different platforms.

+++ spec/std/process_spec.cr
@@ -171,7 +171,7 @@ describe Process do
 
   it "gets the pgid of a process id" do
     process = Process.new("yes")
-    Process.pgid(process.pid).should be_a(Int32)
+    Process.pgid(process.pid).should be_a(Int64)
     process.signal(Signal::KILL)
     Process.pgid.should eq(Process.pgid(Process.pid))
   end
